Added a check to see if the file exists

+++ install.sh
@@ -2,9 +2,18 @@
 
 #
 # A small install script for installing my configuration
-# TODO: Add a check to see if the file exists and diff them in that case
 # TODO: Add questions on which configuration files to install.
+# TODO: Refactor out a function to do this stuff
 #
 
-ln vimrc ../.vimrc
-ln rtorrent.rc ../.rtorrent.rc
+if [ $(ls ~/.vimrc) = 0 ]; then #file exists
+    ln vimrc ../.vimrc
+else
+    echo "The file ~/.vimrc exists, will not install vimrc"
+fi
+
+if [ $(ls ~/.vimrc) = 0 ]; then #file exists
+    ln rtorrent.rc ../.rtorrent.rc
+else
+    echo "The file ~/.vimrc exists, will not install vimrc"
+fi
